feat: Add support for more specific variant types
- add ability to specify truncating mutations, truncating frameshifts
  etc
- split variant file into 2 separate files
